PR-review fixes
