+++ test/functionalTest/cases/1164_v2_update_entities_attr_empty_value/update_entities_attr_empty_value.test
@@ -32,7 +32,7 @@ brokerStart CB
 #
 # 01. POST /v2/entities to create E1/A1=0
 # 02. GET /v2/entities/E1/attrs/A1
-# 03. PUT /v2/entities/E1/attrs/A1/value (empty)
+# 03. PUT /v2/entities/E1/attrs/A1 (empty)
 # 04. GET /v2/entities/E1/attrs/A1, to see A1 empty
 #
